% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_functions.R
\name{assignRain}
\alias{assignRain}
\title{Assign the rain}
\usage{
assignRain(clusterMeans, data, result, emptyDroplets, firstClusters,
  secondClusters, thirdClusters, fourthCluster)
}
\arguments{
\item{clusterMeans}{The cluster centres returned by the respective clustering algorithm.}

\item{data}{The input data. More specifically, a data frame with two dimensions, each dimension representing the intensity for one color.}

\item{result}{A vector of the same length as data containing the cluster number for each event.}

\item{emptyDroplets}{The number of the cluster containing the bottom left population with no marker.}

\item{firstClusters}{The primary clusters according to clusterMeans.}

\item{secondClusters}{The secondary clusters according to clusterMeans.}

\item{thirdClusters}{The tertiary clusters according to clusterMeans.}

\item{fourthCluster}{The quarternary cluster according to clusterMeans.}
}
\value{
The updated data with the rain correctly assigned to its clusters.
}
\description{
Find the rain and assign it based on the distance to vector lines connecting the cluster centres.
}
