% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddPCR.R
\name{createEnsemble}
\alias{createEnsemble}
\title{Create a cluster ensemble.}
\usage{
createEnsemble(dens = NULL, sam = NULL, peaks = NULL, numOfMarkers, file)
}
\arguments{
\item{dens}{The result of the flowDensity algorithm as a CLUE partition.}

\item{sam}{The result of the samSPECTRAL algorithm as a CLUE partition.}

\item{peaks}{The result of the flowPeaks algorithm as a CLUE partition.}

\item{numOfMarkers}{The number of primary clusters that are expected according the experiment set up.}

\item{file}{The input data. More specifically, a data frame with two dimensions, each dimension representing the intensity for one color.}
}
\value{
\item{data}{The original input data minus the removed events (for plotting)}
\item{confidence}{The agreement between the different clustering results in percent. If all algorithms calculated the same result, the clustering is likely to be correct, thus the confidence is high.}
\item{counts}{The droplet count for each cluster.}
}
\description{
Description...
}
\examples{
example...
}
