% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_functions.R
\name{findDeletion}
\alias{findDeletion}
\title{Find missing clusters}
\usage{
findDeletion(clusterMeans, firstClusters, emptyDroplets, numberOfDeletions,
  dimensions)
}
\arguments{
\item{clusterMeans}{The cluster centres returned by the respective clustering algorithm.}

\item{firstClusters}{The primary clusters according to clusterMeans.}

\item{emptyDroplets}{The number of the cluster containing the bottom left population with no marker.}

\item{numberOfDeletions}{The number of clusters that are missing.}

\item{dimensions}{The maximal amplitudes of the data for scaling.}
}
\value{
The numbers of the missing clusters (according to clusterMeans).
}
\description{
Find missing clusters (caused by genomic deletions) based on their positions.
}
