% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_functions.R
\name{findPrimaryClusters}
\alias{findPrimaryClusters}
\title{Find primary clusters using density}
\usage{
findPrimaryClusters(data, clusterMeans, emptyDroplets, remove = 0, dimensions,
  File, f, NumberOfSinglePos = 4)
}
\arguments{
\item{clusterMeans}{The cluster centres returned by the respective clustering algorithm.}

\item{dimensions}{The maximal amplitudes of the data for scaling.}

\item{File}{The input data. More specifically, a data frame with two dimensions, each dimension representing the intensity for one color.}

\item{f}{The data converted into a flow frame.}
}
\value{
The numbers of the primary clusters (according to clusterMeans).
}
\description{
Find the primary clusters based on their density peaks found by flowDensity.
}
