% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_functions.R
\name{findPrimaryClusters_old}
\alias{findPrimaryClusters_old}
\title{Find primary clusters}
\usage{
findPrimaryClusters_old(data, clusterMeans, emptyDroplets, remove = 0,
  dimensions)
}
\arguments{
\item{data}{The input data. More specifically, a data frame with two dimensions, each dimension representing the intensity for one color.}

\item{clusterMeans}{The cluster centres returned by the respective clustering algorithm.}

\item{emptyDroplets}{The number of the cluster containing the bottom left population with no marker.}

\item{remove}{A vector containing the numbers of all clusters that should be excluded from the search.}

\item{dimensions}{The maximal amplitudes of the data for scaling.}
}
\value{
The numbers of the primary clusters (according to clusterMeans).
}
\description{
Find the primary clusters based on their position.
}
