% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_functions.R
\name{findQuaternaryCluster}
\alias{findQuaternaryCluster}
\title{Find quarternary cluster}
\usage{
findQuaternaryCluster(clusterMeans, emptyDroplets, remove, firstClusters,
  secondClusters = 0, thirdClusters = 0)
}
\arguments{
\item{clusterMeans}{The cluster centres returned by the respective clustering algorithm.}

\item{emptyDroplets}{The number of the cluster containing the bottom left population with no marker.}

\item{remove}{A vector containing the numbers of all clusters that should be excluded from the search.}

\item{firstClusters}{The primary clusters according to clusterMeans.}

\item{secondClusters}{The secondary clusters according to clusterMeans.}

\item{thirdClusters}{The tertiary clusters according to clusterMeans.}
}
\value{
The numbers of the quarternary cluster (according to clusterMeans).
}
\description{
Find the quarternary cluster based on its position.
}
