% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_functions.R
\name{findSecondaryClusters}
\alias{findSecondaryClusters}
\title{Find secondary clusters}
\usage{
findSecondaryClusters(firstClusters, clusterMeans, emptyDroplets, remove,
  dimensions, counts)
}
\arguments{
\item{firstClusters}{The primary clusters according to clusterMeans.}

\item{clusterMeans}{The cluster centres returned by the respective clustering algorithm.}

\item{emptyDroplets}{The number of the cluster containing the bottom left population with no marker.}

\item{remove}{A vector containing the numbers of all clusters that should be excluded from the search.}

\item{dimensions}{The maximal amplitudes of the data for scaling.}

\item{counts}{The droplet count for each cluster.}
}
\value{
The cluster numbers of the secondary clusters (according to clusterMeans) and a correction factor
that represents the calculated offset for each primary cluster.
}
\description{
Find the secondary clusters based on the positions of the primary clusters.
}
