% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_functions.R
\name{findTertiaryClusters}
\alias{findTertiaryClusters}
\title{Find tertiary clusters}
\usage{
findTertiaryClusters(emptyDroplets, firstClusters, secondClusters, remove,
  clusterMeans, correctionFactor, dimensions, counts)
}
\arguments{
\item{emptyDroplets}{The number of the cluster containing the bottom left population with no marker.}

\item{firstClusters}{The primary clusters according to clusterMeans.}

\item{secondClusters}{The secondary clusters according to clusterMeans.}

\item{remove}{A vector containing the numbers of all clusters that should be excluded from the search.}

\item{clusterMeans}{The cluster centres returned by the respective clustering algorithm.}

\item{correctionFactor}{A vector that contains a correction factor for each primary cluster based as returned by findSecondaryClusters.}

\item{dimensions}{The maximal amplitudes of the data for scaling.}

\item{counts}{The droplet count for each cluster.}
}
\value{
The cluster numbers of the tertiary clusters (according to clusterMeans).
}
\description{
Find the tertiary clusters based on the positions of the primary and secondary clusters.
}
