% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_functions.R
\name{mergeClusters}
\alias{mergeClusters}
\title{Merge adjacent clusters}
\usage{
mergeClusters(result, clusterMeans, finalResult, remove, p = 12)
}
\arguments{
\item{result}{A vector of the same length as data containing the cluster number for each event}

\item{clusterMeans}{The cluster centres returned by the respective clustering algorithm.}

\item{finalResult}{A vector containing the numbers of all clusters that have been assigned (typically 16 for a 4-plex).}

\item{remove}{A vector containing the numbers of all clusters that should be excluded from the search.}

\item{p}{The paramater that defines how close the clusters need to be in order to be merged. A larger p means a bigger distance is allowed.}
}
\value{
The updated data with merged clusters.
}
\description{
Merge clusters that a close to each other, based on the parameter p.
}
