% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropClust.R
\name{runDropClust}
\alias{runDropClust}
\title{Find the clusters}
\usage{
runDropClust(files, numOfMarkers = 4, sensitivity = 1)
}
\arguments{
\item{numOfMarkers}{The number of primary clusters that are expected according the experiment set up.}

\item{sensitivity}{An integer between 0.1 and 2 determining sensitivity of the initial clustering, e.g. the number of clusters. A higher value means more clusters are being found. Standard is 1.}

\item{file}{The input files. More specifically, csv files with two dimensions, each dimension representing the intensity for one color.}
}
\value{
\item{data}{The original input data minus the removed events (for plotting)}
\item{confidence}{The agreement between the different clustering results in percent. If all algorithms calculated the same result, the clustering is likely to be correct, thus the confidence is high.}
\item{counts}{The droplet count for each cluster.}
}
\description{
Description.
}
\examples{
# Run dropClust
exampleFiles <- list.files(paste0(find.package("dropClust"), "/extdata"), full.names = T)
result <- runDropClust(exampleFiles)
# Plot the results
exportPlots(result, "./results/", "examples")
}
