% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddPCRclust.R
\name{runPeaks}
\alias{runPeaks}
\title{Find the clusters using flowPeaks}
\usage{
runPeaks(file, sensitivity = 1, numOfMarkers, missingClusters = NULL)
}
\arguments{
\item{file}{The input data. More specifically, a data frame with two dimensions, each dimension representing the intensity for one color channel.}

\item{sensitivity}{An integer between 0.1 and 2 determining sensitivity of the initial clustering, e.g. the number of clusters. A higher value means more clusters are being found. Standard is 1.}

\item{numOfMarkers}{The number of primary clusters that are expected according the experiment set up.}

\item{missingClusters}{A vector containing the number of primary clusters, which are missing in this dataset according to the template.}
}
\value{
\item{data}{The original input data minus the removed events (for plotting)}
\item{counts}{The droplet count for each cluster.}
\item{firstClusters}{The position of the primary clusters.}
\item{partition}{The cluster numbers as a CLUE partition (see clue package for more information).}
}
\description{
Find the rain and assign it based on the distance to vector lines connecting the cluster centres.
}
\examples{
# Run the flowPeaks based approach
exampleFiles <- list.files(paste0(find.package("ddPCRclust"), "/extdata"), full.names = TRUE)
file <- read.csv(exampleFiles[3])
peaksResult <- runPeaks(file = file, numOfMarkers = 4)

# Plot the results
library(ggplot2)
p <- ggplot(data = peaksResult$data, mapping = aes(x = Ch2.Amplitude, y = Ch1.Amplitude))
p <- p + geom_point(aes(color = factor(Cluster)), size = .5, na.rm = TRUE) + ggtitle("flowPeaks example")+theme_bw() + theme(legend.position="none")
p

}
