% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropClust.R
\name{exportPlots}
\alias{exportPlots}
\title{Plot the algorithm's results with ggplot2}
\usage{
exportPlots(data, directory, experiment)
}
\arguments{
\item{data}{The result of the dropClust algorithm}

\item{directory}{The parent directory where the plots should saved}

\item{sensitivity}{The name of this experiment. A subfolder in the parent directory will be created with its name to keep things organized.}
}
\description{
A convinience function that takes the results of the droplClust algorithm and plots them using the ggplot2 library and a custom colour palette.
}
\examples{
# Run dropClust
exampleFiles <- list.files(paste0(find.package("dropClust"), "/extdata"), full.names = T)
result <- runDropClust(exampleFiles)
# Plot the results
exportPlots(result, "./results/", "examples")
}
