% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropClust.R
\name{runDropClust}
\alias{runDropClust}
\title{Run the dropClust algorithm}
\usage{
runDropClust(files, numOfMarkers = 4, sensitivity = 1, template = NULL,
  fast = FALSE)
}
\arguments{
\item{files}{The input file(s). More specifically, csv files with two dimensions, each dimension representing one color choannel of the ddPCR reaction.}

\item{numOfMarkers}{The number of primary clusters that are expected according the experiment set up. Can be ignored if a template is provided. 
Else, a vector with length equal to \code{length(files)} should be provided, containing the number of markers used for the respective reaction.}

\item{sensitivity}{An integer between 0.1 and 2 determining sensitivity of the initial clustering, e.g. the number of clusters. A higher value means more clusters are being found. Standard is 1.}

\item{template}{A csv file containing information about the individual ddPCR runs. An example template is provided with this package. For more information, please check the repository on github.}

\item{fast}{Run a simple version of the algorithm that is about 10x faster. For clean data, this can already deliver very good results. In any case useful to get a quick overview over the data.}
}
\value{
\item{results}{The results of the dropClust algorithm. It contains three fields: \cr
\code{data} The original input data minus the removed events (for plotting) 
\code{confidence} The agreement between the different clustering results in percent
If all parts of the algorithm calculated the same result, the clustering is likely to be correct, thus the confidence is high\cr
\code{counts} The droplet count for each cluster
}
\item{annotations}{The metatdata provided in the header of the template. It contains four fields: \cr
\code{Name} The name given to this ddPCR experiment \cr
\code{Ch1} Color channel 1 (usually HEX) \cr
\code{Ch2} Color channel 2 (usually FAM) \cr
\code{descriptions} Additional descriptions about this ddPCR experiment (e.g. date, exprimentor, etc.)
}
\item{template}{A parsed dataframe containing the template, if one was provided.}
}
\description{
This is the main function of this package. It automatically runs the dropClust algorithm on one or multiple csv files containing the raw data from a ddPCR run with up to 4 markers.
}
\examples{
# Run dropClust
exampleFiles <- list.files(paste0(find.package("dropClust"), "/extdata"), full.names = TRUE)
result <- runDropClust(files = exampleFiles[1:8], template = exampleFiles[9])

}
