% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropClust.R
\name{runDropClust}
\alias{runDropClust}
\title{Run the dropClust algorithm}
\usage{
runDropClust(files, numOfMarkers = 4, sensitivity = 1, template = NULL)
}
\arguments{
\item{files}{The input file(s). More specifically, csv files with two dimensions, each dimension representing one color choannel of the ddPCR reaction.}

\item{numOfMarkers}{The number of primary clusters that are expected according the experiment set up. Can be ignored if a template is provided. 
Else, a vector with length equal to \code{length(files)} should be provided, containing the number of markers used for the respective reaction.}

\item{sensitivity}{An integer between 0.1 and 2 determining sensitivity of the initial clustering, e.g. the number of clusters. A higher value means more clusters are being found. Standard is 1.}

\item{template}{A csv file containing information about the individual ddPCR runs. An example template is provided with this package. For more information, please check the repository on github.}
}
\value{
\item{data}{The original input data minus the removed events (for plotting)}
\item{confidence}{The agreement between the different clustering results in percent. If all algorithms calculated the same result, the clustering is likely to be correct, thus the confidence is high.}
\item{counts}{The droplet count for each cluster.}
}
\description{
This is the main function of this package. It automatically runs the dropClust algorithm on one or multiple csv files containing the raw data from a ddPCR run with up to 4 markers.
}
\examples{
# Run dropClust
exampleFiles <- list.files(paste0(find.package("dropClust"), "/extdata"), full.names = TRUE)
result <- runDropClust(files = exampleFiles[1:8], template = exampleFiles[9])

}
